Extract path filtering partial
